Создание единой карты и обновление её скриптов
